agrichain-spoilage/agrichain-spoilage/agri_spoilage/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
autogluon.tabular==1.2
